Add files via upload
The attached files are an implementation of our technique to compute upper and lower bounds on the capacity of unifilar finite-state channels with feedback. The theoretical background of the technique is presented in the following paper:
"Graph-Based Encoders and their Performance for Finite-State Channels with Feedback", by Oron Sabag, Bashar Huleihel and Haim Permuter.